# doc/api.yaml
openapi: 3.0.0
info: 
  description: WASA Text, chat app backend REST API
  title: WASA Text
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Development server
components:
...
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: UUID
      description: >
        Bearer authentication using the UUID identifier returned by POST /session.
        Example: Authorization: Bearer 5c160d51-6f56-4e80-8b9a-6b3f2a1a6c0b
  schemas:
    Id:
      type: string
      description: Unique identifier (alphanumeric + hyphen, 3–16 chars)
      format: uuid
    Uniquename:
      type: string
      description: Unique displayed name (3–16 chars)
      pattern: '^[a-zA-Z0-9-]{3,16}$'
      minLength: 3
      maxLength: 16
...
    Conversation:
      description: conversation where to send and receive messages
      type: object
      required: [id, type, participants, lastMessage, messages]
      properties: 
        id:
          $ref: "#/components/schemas/Id"
        type:
          type: string
          enum: ["individual", "group"]
          description: Specifies wheter the conversation is between two user (individual) or multiple users (group).
          example: "individual"
        participants:
          type: array
          items:
            $ref: "#/components/schemas/User"
          description: List of usernames involved in the conversations
          minItems: 2
          maxItems: 500
        lastMessage:
          $ref: "#/components/schemas/Message"

        messages:
          type: array
          items: 
            $ref: "#/components/schemas/Message"
          description: List of all messages in the conversation, (displayed in reverse chronological order - later to do)
          minItems: 1
          maxItems: 1000
        updatedAt:
          $ref: "#/components/schemas/Timestamp"
...
    ErrorResponse:
      description: Standard error response format
      type: object
      required: [error]
      properties:
        error:
          description: error object
          type: object
          properties:
            code:
              type: integer
              description: HTTP status code
              example: 404
              minimum: 100
              maximum: 599
            message:
              type: string
              description: Human-readable error message
              example: "Resource not found"
              minLength: 1
              maxLength: 200
              pattern: '^[^\r\n]+$'
            details:
              type: string
              description: Additional error details
              example: "The requested conversation was not found"
              minLength: 1
              maxLength: 200
              pattern: '^[^\r\n]+$'
...
    MessageStatusesList:
      type: object
      description: "Wrapper object for the list of message delivery statuses"
      properties:
        statusList:
          type: array
          description: "An array of delivery status entries, one per message"
          items:
            type: object
            required: [messageId, statuses]
            properties:
              messageId:
                $ref: '#/components/schemas/Id'
              statuses:
                type: array
                description: "An array of delivery status entries, one per message"
                items:
                  $ref: '#/components/schemas/StatusEntry'
                minItems: 1
                maxItems: 500
          minItems: 1
          maxItems: 500

      required:
        - statusList
...
security:
  - bearerAuth: [] # use the same name as above
paths:
  /session:
    
    post:
      security: [] # No security
      tags: ["Login"]
      summary: Logs in the user (registers if new)
      description: |
        If the user does not exist, it will be created,
        and an identifier is returned.
        If the user exists, the user identifier is returned.
      operationId: doLogin
      requestBody:
        description: User details
        required: true
        content:
          application/json:
            schema:
              description: Login request payload.
              type: object
              properties:
                username:
                  $ref: "#/components/schemas/Uniquename"
              required:
                - username
      responses:
        '201':
          description: User log-in action successful
          content:
            application/json:
              schema:
                description: object to return toke for authorization
                type: object
                properties:
                  identifier:
                    $ref: '#/components/schemas/Id'
                  username:
                    $ref: "#/components/schemas/Uniquename"
        '400':
          description: Invalid request, missing or invalid username.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /conversations/{id}:
    get:
      tags: ["Conversation Management"]
      summary: Retrieve a conversation
      description: Returns the conversation metadata (participants, lastMessage, etc.).
      operationId: getConversation
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: "#/components/schemas/Id"
          description: ID of the conversation.
      responses:
        "200":
          description: Conversation retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "404":
          description: Conversation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /messages:
    post:
      tags: ["Conversation Management"]
      summary: Sends a message in a conversation
      description: |
        sends a new msg to a specified conversation, either as text, image, or gif
      operationId: sendMessage
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              description: message send payload
              required: [conversationId,contentType]
              properties:
                conversationId:
                  $ref: "#/components/schemas/Id"
                contentType:
                  type: string
                  enum: ["text", "image", "gif"]
                  minLength: 3
                  maxLength: 5
                  description: Type of message content.
                text:
                  type: string
                  description: Text content, if contentType is "text".
                  pattern: '^.*?$'
                  minLength: 1
                  maxLength: 200
                file:
                  description: File for image or GIF content.
                  type: string
                  format: binary
            encoding:
              file:
                contentType: 'image/*'
      responses:
        '201':
          description: Message sent successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: Invalid message content or conversation ID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '401':
          description: Unauthorized access, token required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /messages/{id}:
    delete:
      tags: ["Conversation Management"]
      summary: Deletes a specific message
      description: |
        Allows a user to delete their own message in a conversation. 
        Deletion may remove the message entirely or mark it as "deleted".
      operationId: deleteMessage
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: "#/components/schemas/Id"
          description: ID of the message to delete.
      responses:
        '204':
          description: Message deleted successfully.
        '404':
          description: Message not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '403':
          description: Forbidden, user cannot delete another user’s message.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '401':
          description: Unauthorized access, token required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /messages/{id}/forward:
    post:
      tags: ["Conversation Management"]
      summary: Forwards a message to another conversation
      description: |
        Forwards an existing message to another specified conversation.
      operationId: forwardMessage
      parameters:
        - in: path
          description: id of the formarding msg
          name: id
          required: true
          schema:
            $ref: "#/components/schemas/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: payloed needed to find msg
              type: object
              properties:
                targetConversationId:
                  $ref: "#/components/schemas/Id"
      responses:
        '201':
          description: Message forwarded successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          description: Original message or target conversation not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '401':
          description: Unauthorized access, token required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /messages/{id}/reply:
    post:
      tags: ["Conversation Management"]
      summary: Replies to a specific message
      description: |
        Sends a reply to a specified message in the conversation.
      operationId: replyMessage
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: "#/components/schemas/Id"
          description: ID of the message being replied to.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: payloed needed to find msg
              type: object
              properties:
                text:
                  type: string
                  description: Reply text content.
                  pattern: '^.*?$'
                  minLength: 1
                  maxLength: 200
      responses:
        '201':
          description: Reply sent successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          description: Original message not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '401':
          description: Unauthorized access, token required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /groups/{groupName}/leave:
    post:
      tags: ["Group Management"]
      summary: Allows a member to leave the group
      description: |
        Allows a member to leave the group. This action deletes their reactions but leaves their messages intact.
        They will lose access to the group conversation after leaving.
      operationId: leaveGroup
      parameters:
        - in: path
          name: groupName
          required: true
          schema:
            $ref: "#/components/schemas/Uniquename"
          description: Unique name of the group to leave.
      responses:
        '204':
          description: Member left the group successfully.

        '404':
          description: Group or member not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '401':
          description: Unauthorized access, token required.    
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"     
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse' 
...
  /conversations/{conversationId}/messages/status:
    get:
      tags: ["Conversation Management"]
      summary: Retrieves delivery status of messages in a conversation
      description: |
        Retrieves the delivery status for each message in a conversation, indicating if they have been sent, received, or read (the comparison of time stamp and last activity must implemented later to permit the automation of checkmarks update)
      operationId: getMessageStatuses
      parameters:
        - in: path
          name: conversationId
          required: true
          schema:
             $ref: "#/components/schemas/Id"
          description: conversationId required to get the msg status.
      responses:
        '200':
          description: Delivery statuses retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageStatusesList'

        '404':
          description: Conversation or messages not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '401':
          description: Unauthorized access, token required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'